MultiChain - Page crashes when disconnecting wallet on Liquidity/Swap pages (#3008)
