+++ src/pages/AppBody.tsx
@@ -2,13 +2,14 @@ import React from 'react'
 import { useCookies } from 'react-cookie'
 import styled from 'styled-components/macro'
 
-import { NotAvailablePage } from 'components/NotAvailablePage'
 import { isMobile } from 'react-device-detect'
 import Box from '@mui/material/Box'
 import Portal from '@reach/portal'
 import { CenteredFixed } from 'components/LaunchpadMisc/styled'
 import { NetworkNotAvailable } from 'components/Launchpad/NetworkNotAvailable'
 import { DEFAULT_CHAIN_ID } from 'config'
+import { useWeb3React } from 'hooks/useWeb3React'
+import ConnectWalletCard from 'components/NotAvailablePage/ConnectWalletCard'
 
 export const BodyWrapper = styled(Box)<{
   margin?: string
@@ -80,7 +81,12 @@ export default function AppBody({
   maxWidth?: string
   page?: string
 }) {
+  const { account } = useWeb3React()
   const [cookies] = useCookies(['annoucementsSeen'])
+
+  if (!account) {
+    return <ConnectWalletCard />
+  }
   return (
     <React.Fragment>
       <BodyWrapper
